requirements/dev.txt
-r base.txt

django-debug-toolbar
drf-yasg
